Added comma to separate numbers. Added After tax amount to the results of calculations. Changed the value type for income.

+++ src/main/java/org/liubov/statetaxcalculator/service/TaxCalculatorService.java
@@ -4,13 +4,13 @@ import org.liubov.statetaxcalculator.exception.IncomeTaxCalculatorException;
 
 public interface TaxCalculatorService {
 
-    double calculateStateTax(int year, String state, String filingStatus, int income) throws IncomeTaxCalculatorException;
+    double calculateStateTax(int year, String state, String filingStatus, double income) throws IncomeTaxCalculatorException;
 
-    double calculateFederalTax(int year, String filingStatus, int income) throws IncomeTaxCalculatorException;
+    double calculateFederalTax(int year, String filingStatus, double income) throws IncomeTaxCalculatorException;
 
-    double calculateFicaTax(int year, int income);
+    double calculateFicaTax(int year, double income);
 
-    double calculateAdditionalMedicareTax(int year, String filingStatus, int income);
+    double calculateAdditionalMedicareTax(int year, String filingStatus, double income);
 
-    double calculateEffectiveTaxRate(double tax, int income);
+    double calculateEffectiveTaxRate(double tax, double income);
 }
